Update Product Repository unittest

+++ Product-Microservice/src/main/java/aug/bueno/product/microservice/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface ProductRepository {
 
-    Optional<Product> findById(Integer id);
+    Optional<Product> findById(Long id);
 
     List<Product> findAll();
